+++ configuration.py
@@ -37,11 +37,11 @@ class Configuration:
         self.default_branch = "master"
         self._use_sonarqube = False
         self.sonarqube_persistent = True
-        self.remove_analysis_container_on_analysis_end = True
+        self.sonarqube_keep_analysis_container = False
         self.sonarqube_port = 8085
         self.sonarqube_login = "admin"
         self.sonarqube_password = "admin"
-        self.check_whitespace_changes = True
+        self.ignore_whitespace_changes = True
         self.ignore_remote_repo = False
         self.blame_unseen = True
         self.anonymous_mode = False